requirements.txt
flask-mysqldb
werkzeug.security
os